requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]